feat: duckdb extensions in lambda (#75)
@ceholden I took your good work in
https://github.com/developmentseed/labs-375-stac-geoparquet-backend/pull/60
and reshaped it a bit to use the targeted extension fetching from
https://github.com/stac-utils/stacrs/pull/81. Lambda sizes:

- Before (current **main**): 29.2 MB
- With the **uv** and other tweaks, but no DuckDB extensions: 26.2 MB
- With DuckDB extensions: 93.6 MB

Things are broken, because I haven't wired up the server to actually use
those extensions, but I _think_ this is a proof out of using pre-fetched
DuckDB extensions. Lambda is here:
https://us-west-2.console.aws.amazon.com/lambda/home?region=us-west-2#/functions/stac-fastapi-geoparquet-labs-375-de-lambda8B5974B5-W1CWbXEHRA1Y?tab=code

> [!WARNING]
> When I tried to use the wheels directly, the size got too big. For
now, I'm sticking with a build-in-the-container (which is slow, ~10
minutes on my machine), but I need to dig in to why the build is smaller
than the wheel.