feat: add blog post (#82)
The "results" section is extremely light ... we might need to add some
"what does it mean" text?

I also don't talk about the benefit of having the **stac-geoparquet** in
blob storage so folks can query directly against that (instead of going
through a server layer) or just download the file themselves. @zacdezgeo
do you think that's critical to include, or just a confuser?

In service of
https://github.com/developmentseed/communications/issues/819

---------

Co-authored-by: Zac Deziel <zachary.deziel@gmail.com>